financial_advisor_agent/requirements.txt
bytez==3.0.1
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy==1.26.4
python-dotenv>=1.0.0
plotly>=5.15.0
asyncio>=3.4.3
pydantic
